Refactor DebrisScanner to be Extract
